問題修改 增加GetMemberData輸出欄位 BonusQuery增加流水號 車機對應欄位調整

+++ iRentDBV2/dbo/Stored Procedures/usp_HandleCarStatusByMotor.sql
@@ -63,23 +63,23 @@ CREATE PROCEDURE [dbo].[usp_HandleCarStatusByMotor]
     @deviceRSOC VARCHAR(10),
     @deviceRDistance VARCHAR(10),
     @deviceMBA FLOAT,--
-    @deviceMBAA   INT, --
-    @deviceMBAT_Hi   INT, --
-    @deviceMBAT_Lo   INT, --
+    @deviceMBAA   FLOAT, --
+    @deviceMBAT_Hi   FLOAT, --
+    @deviceMBAT_Lo   FLOAT, --
     @deviceRBA FLOAT,
-    @deviceRBAA   INT, --
-    @deviceRBAT_Hi   INT, --
-    @deviceRBAT_Lo   INT, --
+    @deviceRBAA   FLOAT, --
+    @deviceRBAT_Hi   FLOAT, --
+    @deviceRBAT_Lo   FLOAT, --
     @deviceLBA FLOAT,--
-    @deviceLBAA   INT, --
-    @deviceLBAT_Hi   INT, --
-    @deviceLBAT_Lo   INT, --
-    @deviceTMP    INT, --
-    @deviceCur    INT, --
-    @deviceTPS    INT, --
-    @deviceiVOL   INT, --
-    @deviceErr   INT, --
-    @deviceALT   INT, --
+    @deviceLBAA   FLOAT, --
+    @deviceLBAT_Hi   FLOAT, --
+    @deviceLBAT_Lo   FLOAT, --
+    @deviceTMP    FLOAT, --
+    @deviceCur    FLOAT, --
+    @deviceTPS    FLOAT, --
+    @deviceiVOL   FLOAT, --
+    @deviceErr   FLOAT, --
+    @deviceALT   FLOAT, --
     @deviceGx FLOAT,--
     @deviceGy FLOAT,--
     @deviceGz FLOAT,--
